Network.AWS.MigrationHub.DeleteProgressUpdateStream

Description

Deletes a progress update stream, including all of its tasks, which was previously created as an AWS resource used for access control. This API has the following traits:

The only parameter needed for DeleteProgressUpdateStream is the stream name (same as a CreateProgressUpdateStream call).

The call will return, and a background process will asynchronously be doing the actual delete of the stream and all of its resources (tasks, associated resources, resource attributes, created artifacts).

If the stream takes time to be deleted, it might still show up on a ListProgressUpdateStreams call.

CreateProgressUpdateStream , ImportMigrationTask , NotifyMigrationTaskState , and all Associate[*] APIs realted to the tasks belonging to the stream will throw InvalidInputException if the stream of the same name is in the process of being deleted.

Once the stream and all of its resources are deleted, CreateProgressUpdateStream for a stream of the same name will succeed, and that stream will be an entirely new logical resource (without any resources associated with the old stream).

Creating a Request

deleteProgressUpdateStream

Arguments

:: Text
dpusProgressUpdateStreamName

-> DeleteProgressUpdateStream

Creates a value of DeleteProgressUpdateStream with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dpusDryRun - Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

dpusProgressUpdateStreamName - The name of the ProgressUpdateStream.

data DeleteProgressUpdateStream

See: deleteProgressUpdateStream smart constructor.

Request Lenses

dpusDryRun :: Lens' DeleteProgressUpdateStream (Maybe Bool)

Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

dpusProgressUpdateStreamName :: Lens' DeleteProgressUpdateStream Text

The name of the ProgressUpdateStream.

Destructuring the Response

deleteProgressUpdateStreamResponse

Arguments

:: Int
dpusrsResponseStatus

-> DeleteProgressUpdateStreamResponse

Creates a value of DeleteProgressUpdateStreamResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dpusrsResponseStatus - -- | The response status code.

data DeleteProgressUpdateStreamResponse

See: deleteProgressUpdateStreamResponse smart constructor.

Response Lenses

dpusrsResponseStatus :: Lens' DeleteProgressUpdateStreamResponse Int

- | The response status code.