Network.AWS.MigrationHub.DisassociateCreatedArtifact

Description

Disassociates a created artifact of an AWS resource with a migration task performed by a migration tool that was previously associated. This API has the following traits:

A migration user can call the DisassociateCreatedArtifacts operation to disassociate a created AWS Artifact from a migration task.

The created artifact name must be provided in ARN (Amazon Resource Name) format which will contain information about type and region; for example: arn:aws:ec2:us-east-1:488216288981:image/ami-6d0ba87b .

Examples of the AWS resource behind the created artifact are, AMI's, EC2 instance, or RDS instance, etc.

Creating a Request

disassociateCreatedArtifact

Arguments

:: Text
dcaProgressUpdateStream

-> Text
dcaMigrationTaskName

-> Text
dcaCreatedArtifactName

-> DisassociateCreatedArtifact

Creates a value of DisassociateCreatedArtifact with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcaDryRun - Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

dcaProgressUpdateStream - The name of the ProgressUpdateStream.

dcaMigrationTaskName - Unique identifier that references the migration task to be disassociated with the artifact.

dcaCreatedArtifactName - An ARN of the AWS resource related to the migration (e.g., AMI, EC2 instance, RDS instance, etc.)

data DisassociateCreatedArtifact

See: disassociateCreatedArtifact smart constructor.

Request Lenses

dcaDryRun :: Lens' DisassociateCreatedArtifact (Maybe Bool)

Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

dcaProgressUpdateStream :: Lens' DisassociateCreatedArtifact Text

The name of the ProgressUpdateStream.

dcaMigrationTaskName :: Lens' DisassociateCreatedArtifact Text

Unique identifier that references the migration task to be disassociated with the artifact.

dcaCreatedArtifactName :: Lens' DisassociateCreatedArtifact Text

An ARN of the AWS resource related to the migration (e.g., AMI, EC2 instance, RDS instance, etc.)

Destructuring the Response

disassociateCreatedArtifactResponse

Arguments

:: Int
dcarsResponseStatus

-> DisassociateCreatedArtifactResponse

Creates a value of DisassociateCreatedArtifactResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcarsResponseStatus - -- | The response status code.

data DisassociateCreatedArtifactResponse

See: disassociateCreatedArtifactResponse smart constructor.

Response Lenses

dcarsResponseStatus :: Lens' DisassociateCreatedArtifactResponse Int

- | The response status code.