Network.AWS.MigrationHub.ListCreatedArtifacts

Description

Lists the created artifacts attached to a given migration task in an update stream. This API has the following traits:

Gets the list of the created artifacts while migration is taking place.

Shows the artifacts created by the migration tool that was associated by the AssociateCreatedArtifact API.

Lists created artifacts in a paginated interface.

Creating a Request

listCreatedArtifacts

Arguments

:: Text
lcaProgressUpdateStream

-> Text
lcaMigrationTaskName

-> ListCreatedArtifacts

Creates a value of ListCreatedArtifacts with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lcaNextToken - If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken .

lcaMaxResults - Maximum number of results to be returned per page.

lcaProgressUpdateStream - The name of the ProgressUpdateStream.

lcaMigrationTaskName - Unique identifier that references the migration task.

data ListCreatedArtifacts

See: listCreatedArtifacts smart constructor.

Request Lenses

lcaNextToken :: Lens' ListCreatedArtifacts (Maybe Text)

If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken .

lcaMaxResults :: Lens' ListCreatedArtifacts (Maybe Natural)

Maximum number of results to be returned per page.

lcaProgressUpdateStream :: Lens' ListCreatedArtifacts Text

The name of the ProgressUpdateStream.

lcaMigrationTaskName :: Lens' ListCreatedArtifacts Text

Unique identifier that references the migration task.

Destructuring the Response

listCreatedArtifactsResponse

Arguments

:: Int
lcarsResponseStatus

-> ListCreatedArtifactsResponse

Creates a value of ListCreatedArtifactsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lcarsNextToken - If there are more created artifacts than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

lcarsCreatedArtifactList - List of created artifacts up to the maximum number of results specified in the request.

lcarsResponseStatus - -- | The response status code.

data ListCreatedArtifactsResponse

See: listCreatedArtifactsResponse smart constructor.

Response Lenses

lcarsNextToken :: Lens' ListCreatedArtifactsResponse (Maybe Text)

If there are more created artifacts than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

lcarsCreatedArtifactList :: Lens' ListCreatedArtifactsResponse [CreatedArtifact]

List of created artifacts up to the maximum number of results specified in the request.

lcarsResponseStatus :: Lens' ListCreatedArtifactsResponse Int

- | The response status code.