Network.AWS.MigrationHub.ListDiscoveredResources

Description

Lists discovered resources associated with the given MigrationTask .

Creating a Request

listDiscoveredResources

Arguments

:: Text
ldrProgressUpdateStream

-> Text
ldrMigrationTaskName

-> ListDiscoveredResources

Creates a value of ListDiscoveredResources with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ldrNextToken - If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken .

ldrMaxResults - The maximum number of results returned per page.

ldrProgressUpdateStream - The name of the ProgressUpdateStream.

ldrMigrationTaskName - The name of the MigrationTask.

data ListDiscoveredResources

See: listDiscoveredResources smart constructor.

Request Lenses

ldrNextToken :: Lens' ListDiscoveredResources (Maybe Text)

If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken .

ldrMaxResults :: Lens' ListDiscoveredResources (Maybe Natural)

The maximum number of results returned per page.

ldrProgressUpdateStream :: Lens' ListDiscoveredResources Text

The name of the ProgressUpdateStream.

ldrMigrationTaskName :: Lens' ListDiscoveredResources Text

The name of the MigrationTask.

Destructuring the Response

listDiscoveredResourcesResponse

Arguments

:: Int
ldrrsResponseStatus

-> ListDiscoveredResourcesResponse

Creates a value of ListDiscoveredResourcesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ldrrsDiscoveredResourceList - Returned list of discovered resources associated with the given MigrationTask.

ldrrsNextToken - If there are more discovered resources than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

ldrrsResponseStatus - -- | The response status code.

data ListDiscoveredResourcesResponse

See: listDiscoveredResourcesResponse smart constructor.

Response Lenses

ldrrsDiscoveredResourceList :: Lens' ListDiscoveredResourcesResponse [DiscoveredResource]

Returned list of discovered resources associated with the given MigrationTask.

ldrrsNextToken :: Lens' ListDiscoveredResourcesResponse (Maybe Text)

If there are more discovered resources than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

ldrrsResponseStatus :: Lens' ListDiscoveredResourcesResponse Int

- | The response status code.