Network.AWS.MigrationHub.ListMigrationTasks

Description

Lists all, or filtered by resource name, migration tasks associated with the user account making this call. This API has the following traits:

Can show a summary list of the most recent migration tasks.

Can show a summary list of migration tasks associated with a given discovered resource.

Lists migration tasks in a paginated interface.

Creating a Request

listMigrationTasks :: ListMigrationTasks

Creates a value of ListMigrationTasks with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lmtResourceName - Filter migration tasks by discovered resource name.

lmtNextToken - If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken .

lmtMaxResults - Value to specify how many results are returned per page.

data ListMigrationTasks

See: listMigrationTasks smart constructor.

Request Lenses

lmtResourceName :: Lens' ListMigrationTasks (Maybe Text)

Filter migration tasks by discovered resource name.

lmtNextToken :: Lens' ListMigrationTasks (Maybe Text)

If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken .

lmtMaxResults :: Lens' ListMigrationTasks (Maybe Natural)

Value to specify how many results are returned per page.

Destructuring the Response

listMigrationTasksResponse

Arguments

:: Int
lmtrsResponseStatus

-> ListMigrationTasksResponse

Creates a value of ListMigrationTasksResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lmtrsMigrationTaskSummaryList - Lists the migration task's summary which includes: MigrationTaskName , ProgressPercent , ProgressUpdateStream , Status , and the UpdateDateTime for each task.

lmtrsNextToken - If there are more migration tasks than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

lmtrsResponseStatus - -- | The response status code.

data ListMigrationTasksResponse

See: listMigrationTasksResponse smart constructor.

Response Lenses

lmtrsMigrationTaskSummaryList :: Lens' ListMigrationTasksResponse [MigrationTaskSummary]

Lists the migration task's summary which includes: MigrationTaskName , ProgressPercent , ProgressUpdateStream , Status , and the UpdateDateTime for each task.

lmtrsNextToken :: Lens' ListMigrationTasksResponse (Maybe Text)

If there are more migration tasks than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

lmtrsResponseStatus :: Lens' ListMigrationTasksResponse Int

- | The response status code.