Network.AWS.MigrationHub.PutResourceAttributes

Description

Provides identifying details of the resource being migrated so that it can be associated in the Application Discovery Service (ADS)'s repository. This association occurs asynchronously after PutResourceAttributes returns.

Important: Keep in mind that subsequent calls to PutResourceAttributes will override previously stored attributes. For example, if it is first called with a MAC address, but later, it is desired to add an IP address, it will then be required to call it with both the IP and MAC addresses to prevent overiding the MAC address.

Creating a Request

putResourceAttributes

Arguments

:: Text
praProgressUpdateStream

-> Text
praMigrationTaskName

-> NonEmpty ResourceAttribute
praResourceAttributeList

-> PutResourceAttributes

Creates a value of PutResourceAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

praDryRun - Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

praProgressUpdateStream - The name of the ProgressUpdateStream.

praMigrationTaskName - Unique identifier that references the migration task.

praResourceAttributeList - Information about the resource that is being migrated. This data will be used to map the task to a resource in the Application Discovery Service (ADS)'s repository.

data PutResourceAttributes

See: putResourceAttributes smart constructor.

Request Lenses

praDryRun :: Lens' PutResourceAttributes (Maybe Bool)

Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

praProgressUpdateStream :: Lens' PutResourceAttributes Text

The name of the ProgressUpdateStream.

praMigrationTaskName :: Lens' PutResourceAttributes Text

Unique identifier that references the migration task.

praResourceAttributeList :: Lens' PutResourceAttributes (NonEmpty ResourceAttribute)

Information about the resource that is being migrated. This data will be used to map the task to a resource in the Application Discovery Service (ADS)'s repository.

Destructuring the Response

putResourceAttributesResponse

Arguments

:: Int
prarsResponseStatus

-> PutResourceAttributesResponse

Creates a value of PutResourceAttributesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

prarsResponseStatus - -- | The response status code.

data PutResourceAttributesResponse

See: putResourceAttributesResponse smart constructor.

Response Lenses

prarsResponseStatus :: Lens' PutResourceAttributesResponse Int

- | The response status code.