Network.AWS.MigrationHub.Types

Description

Service Configuration

migrationHub :: Service

API version 2017-05-31 of the Amazon Migration Hub SDK configuration.

Errors

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised when the account making the call is not whitelisted or there are other authentication errors.

_DryRunOperation :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised to indicate a successfully authorized action when the DryRun flag is set to "true".

_PolicyErrorException :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised when there are problems accessing ADS (Application Discovery Service); most likely due to a misconfigured policy or the ADSCaller role is missing or not configured correctly.

_InternalServerError :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised when there is an internal, configuration, or dependency error encountered.

_InvalidInputException :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised when the provided input violates a policy constraint or is entered in the wrong format or data type.

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised when the service encounters throttled communication with upstream dependencies or is overloaded with requests.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised when the request references a resource (ADS configuration, update stream, migration task, etc.) that does not exist in ADS (Application Discovery Service) or in Migration Hub's repository.

_UnauthorizedOperation :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised to indicate a request was not authorized when the DryRun flag is set to "true".

ApplicationStatus

data ApplicationStatus

Constructors

ASCompleted

ASInProgress

ASNotStarted

MigrationStatus

data MigrationStatus

Constructors

Completed

Failed

InProgress

NotStarted

ResourceAttributeType

data ResourceAttributeType

Constructors

BiosId

Fqdn

IPV4Address

IPV6Address

Label

MACAddress

MotherboardSerialNumber

VMManagedObjectReference

VMManagerId

VMName

VMPath

CreatedArtifact

data CreatedArtifact

An ARN of the AWS cloud resource target receiving the migration (e.g., AMI, EC2 instance, RDS instance, etc.).

See: createdArtifact smart constructor.

createdArtifact

Arguments

:: Text
caName

-> CreatedArtifact

Creates a value of CreatedArtifact with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

caDescription - A description that can be free-form text to record additional detail about the artifact for clarity or for later reference.

caName - An ARN that uniquely identifies the result of a migration task.

caDescription :: Lens' CreatedArtifact (Maybe Text)

A description that can be free-form text to record additional detail about the artifact for clarity or for later reference.

caName :: Lens' CreatedArtifact Text

An ARN that uniquely identifies the result of a migration task.

DiscoveredResource

data DiscoveredResource

Object representing the on-premises resource being migrated.

See: discoveredResource smart constructor.

discoveredResource

Arguments

:: Text
drConfigurationId

-> DiscoveredResource

Creates a value of DiscoveredResource with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drDescription - A description that can be free-form text to record additional detail about the discovered resource for clarity or later reference.

drConfigurationId - The configurationId in ADS that uniquely identifies the on-premise resource.

drDescription :: Lens' DiscoveredResource (Maybe Text)

A description that can be free-form text to record additional detail about the discovered resource for clarity or later reference.

drConfigurationId :: Lens' DiscoveredResource Text

The configurationId in ADS that uniquely identifies the on-premise resource.

MigrationTask

data MigrationTask

Represents a migration task in a migration tool.

See: migrationTask smart constructor.

migrationTask :: MigrationTask

Creates a value of MigrationTask with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mtUpdateDateTime - The timestamp when the task was gathered.

mtResourceAttributeList -

mtTask - Task object encapsulating task information.

mtProgressUpdateStream - A name that identifies the vendor of the migration tool being used.

mtMigrationTaskName - Unique identifier that references the migration task.

mtUpdateDateTime :: Lens' MigrationTask (Maybe UTCTime)

The timestamp when the task was gathered.

mtResourceAttributeList :: Lens' MigrationTask [ResourceAttribute]

mtTask :: Lens' MigrationTask (Maybe Task)

Task object encapsulating task information.

mtProgressUpdateStream :: Lens' MigrationTask (Maybe Text)

A name that identifies the vendor of the migration tool being used.

mtMigrationTaskName :: Lens' MigrationTask (Maybe Text)

Unique identifier that references the migration task.

MigrationTaskSummary

data MigrationTaskSummary

MigrationTaskSummary includes MigrationTaskName , ProgressPercent , ProgressUpdateStream , Status , and UpdateDateTime for each task.

See: migrationTaskSummary smart constructor.

migrationTaskSummary :: MigrationTaskSummary

Creates a value of MigrationTaskSummary with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mtsStatus - Status of the task.

mtsUpdateDateTime - The timestamp when the task was gathered.

mtsProgressPercent -

mtsStatusDetail - Detail information of what is being done within the overall status state.

mtsProgressUpdateStream - An AWS resource used for access control. It should uniquely identify the migration tool as it is used for all updates made by the tool.

mtsMigrationTaskName - Unique identifier that references the migration task.

mtsStatus :: Lens' MigrationTaskSummary (Maybe MigrationStatus)

Status of the task.

mtsUpdateDateTime :: Lens' MigrationTaskSummary (Maybe UTCTime)

The timestamp when the task was gathered.

mtsProgressPercent :: Lens' MigrationTaskSummary (Maybe Natural)

mtsStatusDetail :: Lens' MigrationTaskSummary (Maybe Text)

Detail information of what is being done within the overall status state.

mtsProgressUpdateStream :: Lens' MigrationTaskSummary (Maybe Text)

An AWS resource used for access control. It should uniquely identify the migration tool as it is used for all updates made by the tool.

mtsMigrationTaskName :: Lens' MigrationTaskSummary (Maybe Text)

Unique identifier that references the migration task.

ProgressUpdateStreamSummary

data ProgressUpdateStreamSummary

Summary of the AWS resource used for access control that is implicitly linked to your AWS account.

See: progressUpdateStreamSummary smart constructor.

progressUpdateStreamSummary :: ProgressUpdateStreamSummary

Creates a value of ProgressUpdateStreamSummary with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pussProgressUpdateStreamName - The name of the ProgressUpdateStream.

pussProgressUpdateStreamName :: Lens' ProgressUpdateStreamSummary (Maybe Text)

The name of the ProgressUpdateStream.

ResourceAttribute

data ResourceAttribute

Attribute associated with a resource.

See: resourceAttribute smart constructor.

resourceAttribute

Arguments

:: ResourceAttributeType
raType

-> Text
raValue

-> ResourceAttribute

Creates a value of ResourceAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

raType - Type of resource.

raValue - Value of the resource type.

raType :: Lens' ResourceAttribute ResourceAttributeType

Type of resource.

raValue :: Lens' ResourceAttribute Text

Value of the resource type.

Task

data Task

Task object encapsulating task information.

See: task smart constructor.

task

Arguments

:: MigrationStatus
tStatus

-> Task

Creates a value of Task with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tProgressPercent - Indication of the percentage completion of the task.

tStatusDetail - Details of task status as notified by a migration tool. A tool might use this field to provide clarifying information about the status that is unique to that tool or that explains an error state.

tStatus - Status of the task - Not Started, In-Progress, Complete.

tProgressPercent :: Lens' Task (Maybe Natural)

Indication of the percentage completion of the task.

tStatusDetail :: Lens' Task (Maybe Text)

Details of task status as notified by a migration tool. A tool might use this field to provide clarifying information about the status that is unique to that tool or that explains an error state.

tStatus :: Lens' Task MigrationStatus

Status of the task - Not Started, In-Progress, Complete.